Saga implementation

+++ src/main/java/com/tokentrackr/asset_service/service/interfaces/PriceCacheService.java
@@ -1,5 +1,6 @@
 package com.tokentrackr.asset_service.service.interfaces;
 
+import java.math.BigDecimal;
 import java.util.Map;
 import java.util.Set;
 
@@ -8,17 +9,17 @@ public interface PriceCacheService {
     /**
      * Store price for a symbol in Redis cache
      */
-    void cachePrice(String symbol, Double price);
+    void cachePrice(String symbol, BigDecimal price);
 
     /**
      * Get price for a symbol from Redis cache
      */
-    Double getPrice(String symbol);
+    BigDecimal getPrice(String symbol);
 
     /**
      * Get prices for multiple symbols
      */
-    Map<String, Double> getPrices(Set<String> symbols);
+    Map<String, BigDecimal> getPrices(Set<String> symbols);
 
     /**
      * Check if price exists in cache
